+++ perf/tuple.cc
@@ -4,7 +4,6 @@
 #include "memtx_engine.h"
 #include <allocator.h>
 
-#include <iostream>
 #include <benchmark/benchmark.h>
 
 const size_t NUM_TEST_TUPLES = 4096;
@@ -435,24 +434,4 @@ BENCHMARK(tuple_tuple_compare_hint);
 
 BENCHMARK_MAIN();
 
-static void
-show_warning_if_debug()
-{
-#ifndef NDEBUG
-	std::cerr << "#######################################################\n"
-		  << "#######################################################\n"
-		  << "#######################################################\n"
-		  << "###                                                 ###\n"
-		  << "###                    WARNING!                     ###\n"
-		  << "###   The performance test is run in debug build!   ###\n"
-		  << "###   Test results are definitely inappropriate!    ###\n"
-		  << "###                                                 ###\n"
-		  << "#######################################################\n"
-		  << "#######################################################\n"
-		  << "#######################################################\n";
-#endif // #ifndef NDEBUG
-}
-
-struct DebugWarning {
-	DebugWarning() { show_warning_if_debug(); }
-} debug_warning;
+#include "debug_warning.h"
